Add distribution Change float to double Change histogram builder

+++ src/by/bsuir/verkpavel/saimmod/distributions/GammaDistribution.java
@@ -1,11 +1,11 @@
 package by.bsuir.verkpavel.saimmod.distributions;
 
 public class GammaDistribution extends BaseDistribution {
-	private float lambda;
+	private double lambda;
 	private int setLength;
 	private BaseDistribution generator;
 	
-	public GammaDistribution(float lambda, int  n, int length) {
+	public GammaDistribution(double lambda, int  n, int length) {
 		super(length);
 		this.lambda = lambda;
 		this.setLength = n;
@@ -16,11 +16,11 @@ public class GammaDistribution extends BaseDistribution {
 	@Override
 	public void build() {
 		for(int i=0; i<this.count; i++){
-			float sum =1;
+			double sum =1;
 			for(int j=0; j<this.setLength; j++){
 				sum*=generator.items.get(this.setLength*i+j);
 			}
-			items.add((float) (-1/this.lambda*Math.log(sum)));
+			items.add((double) (-1/this.lambda*Math.log(sum)));
 		}
 
 	}
